+++ SDK/backendlessAPI/Classes/weborb/Util/Common/CrowdNode.m
@@ -77,7 +77,7 @@
 }
 
 -(int)count {
-	return node.count;
+	return (int)node.count;
 }
 
 -(NSArray *)keys {
